uv-resolver: rewrite forking to use marker disjointness
The previous way we did forking was merely by looking at whether there were
duplicate dependency specifications (with respect to package name), and if
their version constraints were different, we forked. But this is
incorrect (and was known to be, it was a shortcut). We only want to fork
when we can prove that the resolution produced will never result in
installing two different versions of the same package. That in turn can
only happen when there are marker expressions attached to conflicting
dependency specifications that never overlap.

This does slightly change the test output, but only the ordering of
packages. This might be because of the fork refactoring. We may want to
investigate making the test output more stable, but we punt on that for
now.