Don't panic with invalid wheel source
Remove the panic when there is an invalid wheel source, instead surface the error. This error can only occur when manually editing the lock file, but since it's an external file, we should error and not panic.

This change is helpful since the method needs to be able to error for relative path support.