Move adding dependencies for versions into dedicated method
To support diverging urls, we have to check urls when adding dependencies (after forking). To prepare for this, i've moved adding dependencies for the current version to `SolveState::add_package_version_dependencies` and removed the duplication when checking for self-dependencies.

This changed is joined with a change in pubgrub (https://github.com/astral-sh/pubgrub/pull/27) that simplifies the same code path.