uv/tests: update snapshots for sdist omission
This updates all of the test snapshots where `sdist` was
strictly redundant and could be removed.

Note that there is one test failure whose snapshot I didn't
update: one where there is a direct URL dependency. In this
case, the sdist entry isn't strictly redundant, as it includes
a hash that isn't present in the source. We'll deal with that
in a subsequent commit.